This file was fixed for the bug.
git-svn-id: https://tmp.openrtm.org/svn/OpenRTM-aist-Java/branches/RELENG_0_4_1@18 9a6ac7ee-f497-465d-b832-7029f4913c1d

+++ jp.go.aist.rtm.RTC/src/RTMExamples/GUIIn/GUIInImpl.java
@@ -96,13 +96,13 @@ public class GUIInImpl  extends DataFlowComponentBase {
     @Override
     protected ReturnCode_t onExecute(int ec_id) {
         if( guiIn.getSliderValue().isEnabled()) {
-            m_out1_val.data = guiIn.getSliderValue().getData();
+            m_out1_val.data = guiIn.getSliderValue().getData().intValue();
         }
         if( guiIn.getTextValue().isEnabled()) {
             m_out2_val.data = guiIn.getTextValue().getData();
         }
         if( guiIn.getSpinnerValue().isEnabled()) {
-            m_out3_val.data = guiIn.getSpinnerValue().getData();
+            m_out3_val.data = guiIn.getSpinnerValue().getData().doubleValue();
         }
         m_out1Out.write();
         m_out2Out.write();
